+++ src/MovieCard/MovieCard.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import '../CSS/MovieCard.css'
+import './MovieCard.css'
 
 
 function MovieCard(props) {
